controllight_update: int bit + (u8) cast drops redundant clrlwi before cntlzw
bit==0 on a u8 re-masked before cntlzw; int bit holding the masked
GameBit_Get return reuses the value directly (matches target cntlzw r,bit).
Remaining gap is pure #108 saved-reg coloring.

+++ src/main/dll/LGT/dll_02AC_lgtcontrollight.c
@@ -79,7 +79,7 @@ void controllight_update(int obj)
 {
 
     ControlLightState* state = ((GameObject*)obj)->extra;
-    u8 bit = GameBit_Get(state->gameBit);
+    int bit = (u8)GameBit_Get(state->gameBit);
 
     if (bit != state->lastBit)
     {
